+++ src/main/java/persistence/entity/EntityPersister.java
@@ -2,10 +2,12 @@ package persistence.entity;
 
 import jakarta.persistence.Id;
 import persistence.sql.ddl.exception.IdAnnotationMissingException;
-import persistence.sql.dml.*;
+import persistence.sql.dml.DeleteQueryBuilder;
+import persistence.sql.dml.InsertQueryBuilder;
+import persistence.sql.dml.UpdateQueryBuilder;
+import persistence.sql.dml.WhereBuilder;
 import persistence.sql.mapping.ColumnData;
 import persistence.sql.mapping.Columns;
-import jdbc.JdbcTemplate;
 import persistence.sql.mapping.TableData;
 
 import java.lang.reflect.Field;
@@ -13,69 +15,8 @@ import java.util.Arrays;
 
 import static persistence.sql.dml.BooleanExpression.eq;
 
-public class EntityPersister {
-    private final GeneratedIdObtainStrategy generatedIdObtainStrategy;
-    private final JdbcTemplate jdbcTemplate;
-
-    public EntityPersister(GeneratedIdObtainStrategy generatedIdObtainStrategy, JdbcTemplate jdbcTemplate) {
-        this.generatedIdObtainStrategy = generatedIdObtainStrategy;
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public boolean update(Object entity) {
-        TableData table = TableData.from(entity.getClass());
-        Columns columns = Columns.createColumnsWithValue(entity);
-        ColumnData keyColumn = columns.getKeyColumn();
-
-        if(keyColumn.getValue() == null) {
-            return false;
-        }
-
-        UpdateQueryBuilder updateQueryBuilder = new UpdateQueryBuilder(table, columns);
-        WhereBuilder whereBuilder = new WhereBuilder();
-        whereBuilder.and(eq(keyColumn.getName(), keyColumn.getValue()));
-
-        jdbcTemplate.execute(updateQueryBuilder.build(entity, whereBuilder));
-
-        return true;
-    }
-
-    public void insert(Object entity) {
-        Class<?> clazz = entity.getClass();
-        InsertQueryBuilder insertQueryBuilder = new InsertQueryBuilder(clazz);
-
-        jdbcTemplate.execute(insertQueryBuilder.build(entity));
-
-        setIdToEntity(entity, clazz);
-    }
-
-    private void setIdToEntity(Object entity, Class<?> clazz) {
-        Object generatedId = jdbcTemplate.queryForObject(
-                generatedIdObtainStrategy.getQueryString(),
-                generatedIdObtainStrategy.getRowMapper()
-        );
-
-        Field idField = Arrays.stream(clazz.getDeclaredFields())
-                .filter(field -> field.isAnnotationPresent(Id.class))
-                .findFirst()
-                .orElseThrow();
-        idField.setAccessible(true);
-
-        try {
-            idField.set(entity, generatedId);
-        } catch (IllegalAccessException e) {
-            throw new IdAnnotationMissingException();
-        }
-    }
-
-    public void delete(Object entity) {
-        Class<?> clazz = entity.getClass();
-        ColumnData idColumn = Columns.createColumnsWithValue(entity).getKeyColumn();
-
-        DeleteQueryBuilder deleteQueryBuilder = new DeleteQueryBuilder(clazz);
-        WhereBuilder builder = new WhereBuilder();
-        builder.and(eq(idColumn.getName(), idColumn.getValue()));
-
-        jdbcTemplate.execute(deleteQueryBuilder.build(builder));
-    }
+public interface EntityPersister {
+    boolean update(Object entity);
+    void insert(Object entity);
+    void delete(Object entity);
 }
